requirements.txt
fastapi
uvicorn
celery
redis
pymongo
requests
python-dotenv
jinja2